sa_range written

+++ DS_in_C/utilities/array_utils/array_utils.h
@@ -1,7 +1,7 @@
 #ifndef ARRAY_UTILS_H
 #define ARRAY_UTILS_H
 
-void buildArray(int* arr, int arr_len, int lower, int jump);
+void buildArray(int* arr, size_t arr_len, int lower, int jump);
 void printArray(int* arr, size_t arr_len);
 void reverse(int* arr, size_t arr_len);
